Feat: CopyURL 공통 컴포넌트로 분리, 테스트 결과 페이지에 재사용
